No longer wrap and run ct in Docker container
With this major change, the action no longer wraps and runs ct in a Docker
container but simply installs it directly on the host. Default config
files that are part of the distro and Yamale and yamllint are installed as
well.

With this change, ct is directly exposed to the user with all options
directly available. It will be much easier now to integrate custom tools
using the new feature to run additional commands introduced in #283.

Signed-off-by: Reinhard Nägele <unguiculus@gmail.com>